pyproject.toml
[project]
name = "blog-ihribernik-ar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.4.2",
    "psycopg2-binary>=2.9.6",
    "pydantic-settings>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.3.0",
    "isort>=6.0.1",
    "mypy>=1.17.1",
    "pylance>=0.35.0",
    "pylint>=3.3.8",
    "pytest==8.4.1",
    "types-passlib>=1.7.7.20250602",
    "types-python-jose>=3.5.0.20250531",
]

[tool.black]
line-length = 88
target-version = ["py313"]
include = '\.pyi?$'
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
^/migrations/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip = ["migrations"]
known_first_party = ["app"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
python_version = "3.13"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_optional = true
exclude = ['^migrations/']

[[tool.mypy.overrides]]
module = ["sqlalchemy.*", "fastapi.*", "pydantic.*", "dotenv.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.4"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
